Adds tests
- Adds tests for Forecast model.
- Updates tests for Current model.
